pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phasephyto"
version = "0.1.0"
description = "Physics-Informed Differentiable Phase Congruency for Cross-Domain Botanical Visual Recognition"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "timm>=0.9.12",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "opencv-python-headless>=4.8.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
logging = [
    "wandb>=0.16.0",
]
all = ["phasephyto[dev,logging]"]

[tool.setuptools.packages.find]
where = ["."]
include = ["phasephyto*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disable_error_code = ["import-untyped"]
